#1 Move basic blueprint v2 structs into new library
