+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/cloudogu/blueprint-lib
+module github.com/cloudogu/k8s-blueprint-lib
 
 go 1.23.4
