+++ json/blueprintMaskV1/serializer.go
