+++ json/bpcore/core.go
