+++ json/entities/targetComponent.go
