+++ json/entities/targetDogu.go
