--- v1/blueprint_test.go
@@ -1,107 +1,0 @@
-package v1
-
-import (
-	"encoding/json"
-	"github.com/cloudogu/blueprint-lib/bpcore"
-	"github.com/stretchr/testify/require"
-	"reflect"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestTargetState_String(t *testing.T) {
-	tests := []struct {
-		name  string
-		state bpcore.TargetState
-		want  string
-	}{
-		{
-			"String() map enum to string",
-			bpcore.TargetStatePresent,
-			"present",
-		},
-		{
-			"String() map enum to string",
-			bpcore.TargetStateAbsent,
-			"absent",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.state.String(); got != tt.want {
-				t.Errorf("TargetState.String() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestTargetState_MarshalJSON(t *testing.T) {
-	tests := []struct {
-		name    string
-		state   bpcore.TargetState
-		want    []byte
-		wantErr bool
-	}{
-		{
-			"MarshalJSON to bytes",
-			bpcore.TargetStatePresent,
-			[]byte(`"present"`),
-			false,
-		},
-		{
-			"MarshalJSON to bytes",
-			bpcore.TargetStateAbsent,
-			[]byte(`"absent"`),
-			false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := tt.state.MarshalJSON()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("TargetState.MarshalJSON() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("TargetState.MarshalJSON() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestTargetState_UnmarshalJSON_secondValue(t *testing.T) {
-	jsonBlob := []byte(`"absent"`)
-	var sut bpcore.TargetState
-
-	err := json.Unmarshal(jsonBlob, &sut)
-
-	require.NoError(t, err)
-	assert.EqualValues(t, bpcore.TargetStateAbsent, sut)
-}
-
-func TestTargetState_UnmarshalJSON_firstValue(t *testing.T) {
-	jsonBlob := []byte(`"present"`)
-	var sut bpcore.TargetState
-	err := json.Unmarshal(jsonBlob, &sut)
-
-	require.NoError(t, err)
-	assert.EqualValues(t, bpcore.TargetStatePresent, sut)
-}
-
-func TestTargetState_UnmarshalJSON_unknownValueParsesToFirstState(t *testing.T) {
-	jsonBlob := []byte(`"test"`)
-	var sut bpcore.TargetState
-	err := json.Unmarshal(jsonBlob, &sut)
-
-	require.NoError(t, err)
-	assert.EqualValues(t, bpcore.TargetStatePresent, sut)
-}
-
-func TestTargetState_UnmarshalJSON_error(t *testing.T) {
-	jsonBlob := []byte("test")
-	var sut bpcore.TargetState
-	err := json.Unmarshal(jsonBlob, &sut)
-
-	assert.Error(t, err)
-}
